Adding AuthzUtilTest and Refactored OAuth2AuthzEndpointTest
